FIX(pages)(Faq): wrong copy icon tooltip on click
introduce Section component
